APPNG-2128 prevent running site from being deleted

+++ appng-appngizer/src/main/java/org/appng/appngizer/controller/SiteController.java
@@ -21,12 +21,16 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.ServletContext;
+
 import org.apache.commons.io.FileUtils;
 import org.appng.api.BusinessException;
 import org.appng.api.Environment;
+import org.appng.api.RequestUtil;
 import org.appng.api.SiteProperties;
 import org.appng.api.messaging.Messaging;
 import org.appng.api.messaging.Sender;
+import org.appng.api.model.Site.SiteState;
 import org.appng.api.support.FieldProcessorImpl;
 import org.appng.api.support.environment.DefaultEnvironment;
 import org.appng.appngizer.model.Link;
@@ -152,10 +156,17 @@ public class SiteController extends ControllerBase {
 	}
 
 	@DeleteMapping(value = "/site/{name}")
-	public ResponseEntity<Void> deleteSite(@PathVariable("name") String name) throws BusinessException {
+	public ResponseEntity<Void> deleteSite(@PathVariable("name") String name, ServletContext context)
+			throws BusinessException {
 		SiteImpl currentSite = getSiteByName(name);
 		if (null == currentSite) {
 			return notFound();
+		}
+		Environment environment = DefaultEnvironment.get(context);
+		org.appng.api.model.Site site = RequestUtil.getSiteByName(environment, name);
+		if (site != null && (site.getState() == SiteState.STARTING || site.getState() == SiteState.STARTED
+				|| site.getState() == SiteState.STOPPING)) {
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 		}
 		getCoreService().deleteSite(name, new FieldProcessorImpl("delete-site"));
 		HttpHeaders headers = new HttpHeaders();
